requirements.txt
pdf-craft-cloud==0.1.3
dataclasses-json==0.6.7
